requirements.txt
flask
scikit-learn
pandas
numpy
deep-translator
pytesseract
pillow
requests
beautifulsoup4
openpyxl
